making register work better
