+++ src/App.js
@@ -1,11 +1,12 @@
 import "./App.css";
 import LoginForm from "./LoginForm";
-import { RegisterForm } from "./RegisterForm";
+import { RegisterFormFirst } from "./registerForm/RegisterForm.js";
+import { RegisterFormExtras } from "./registerForm/RegisterForm.js";
 
 function App() {
   return (
     <div>
-      <RegisterForm />
+      <RegisterFormFirst />
     </div>
   );
 }
