--- src/RegisterForm.css
@@ -1,32 +1,0 @@
-.center {
-    margin: auto;
-    width: 50%;
-}
-
-.textbox {
-    width: 100%;
-    height: 1%;
-    padding: 2% 2%;
-    margin: 1% 0%;
-    box-sizing: border-box;
-}
-
-.form-area {
-    width: 40%;
-    text-align: center;
-    margin-top: 3%;
-}
-
-.right-align {
-    float: right
-}
-
-.login-button {
-    margin: 5%;
-    float: right
-}
-
-::placeholder {
-    color: #8D8D8D;
-    font-size: 120%;
-}
